feat: Cuba product-grid UI overhaul — orders, cart, wishlist
- Rewrote create.blade.php with exact Cuba product-grid structure:
  exact product-box/product-hover/product-details classes,
  working filter sidebar with checkbox_animated + price range,
  live search, sort dropdown, grid/list/column toggle,
  data-bs-toggle modals (no bootstrap.Modal JS), compare bar,
  sticky selected-plan bar, always-visible select button
- Fixed bootstrap not defined error by using data-bs-toggle attributes
- Rewrote index.blade.php (order history) to match order-history.html:
  date/status filter card, order-history-table with all columns,
  view+invoice action buttons, pagination
- Rewrote show.blade.php (order detail) to match order-details.html:
  status timeline (5 steps), order items table, summary sidebar,
  customer details card, invoice link
- Added CartController + panel/cart/index.blade.php (Cuba cart.html style)
- Added WishlistController + panel/wishlist/index.blade.php (wishlist.html)
- Added cart/wishlist routes (kosik, oblibene)
- Fixed billing.proforma_validity_days int cast (was string from env)

+++ config/billing.php
@@ -32,8 +32,8 @@ return [
     | Invoicing
     |--------------------------------------------------------------------------
     */
-    'invoice_due_days'        => env('BILLING_INVOICE_DUE_DAYS', 14),
-    'proforma_validity_days'  => env('BILLING_PROFORMA_VALIDITY_DAYS', 10),
+    'invoice_due_days'        => (int) env('BILLING_INVOICE_DUE_DAYS', 14),
+    'proforma_validity_days'  => (int) env('BILLING_PROFORMA_VALIDITY_DAYS', 10),
 
     'supplier' => [
         'name'    => env('BILLING_COMPANY_NAME', 'Onhost.cz s.r.o.'),
